+++ web/src/pages/Explore/HoneyComb/index.tsx
@@ -143,7 +143,6 @@ function HoneyComb({ router, selectedNetwork }) {
             />
           </ShowDesktopOnly>
         </div>
-          <DisplayInstructions/>
           <HexagonExploreMap
             exploreSettings={exploreSettings}
             h3TypeDensityHexes={h3TypeDensityHexes}
@@ -643,25 +642,3 @@ const orderBy = (buttons, orderBy, center) => {
   return buttons;
 };
 
-
-function DisplayInstructions() {
-  const loggedInUser = useStore(
-    store,
-    (state: GlobalState) => state.loggedInUser,
-    false,
-  );
-  const showInstructions = useStore(
-    store,
-    (state: GlobalState) => state.explore.map.showInstructions,
-    false,
-  );
-  return (
-    <>
-      {(showInstructions && !loggedInUser ) && (
-        <div className="search-map__instructions">
-          {t('explore.displayInstructions')}
-        </div>
-      )}
-    </>
-  );
-}
